+++ CHANGELOG.md
@@ -1,16 +1,27 @@
-# 2.1.1 - 2017-14-04
+# 3.0.0 - 2017-11-08
 
-* Remove `webpack` build
+* Removed: `webpack` build
+* Added: CommonJS, UMD and ES6 build
+* Added: tests
+* Added: autocompletionRequest prop
+* Removed: suggestRadius prop -> autocompletionRequest prop
+* Removed: suggestTypes prop -> autocompletionRequest prop
+* Removed: search prop -> autocompletionRequest prop
+* Removed: suggestComponentRestrictions prop -> autocompletionRequest prop
+* Updated: renderSuggest prop -> customRender prop
 
+# 2.1.1 - 2017-04-14
+
+* Removed: `webpack` build
 
 # 2.0.0 - 2016-12-11
 
-* Add : `textNoResults` prop
-* Add : `renderSuggest` prop
+* Add: `textNoResults` prop
+* Add: `renderSuggest` prop
 
 # 1.0.1 - 2015-09-11
 
-* Add : libraryTarget into Webpack config
+* Add: libraryTarget into Webpack config
 
 # 1.0.0 - 2015-09-09